read-pdf/requirements.txt
PyPDF2
pdfplumber
pdfminer.six
